Restructured internals of project, splitting up algorithms.cpp
Also added a pointer to a transposition table as a member of the Searcher class to avoid having that hard-coded to a global variable.

Removed ability to turn of killer heuristic and transposition tables through macros. The conditional compilation of these features was proving a maintenance burden, and since they are both fairly well-proven by now, these macro switches were removed. Negascout can still be disabled though, and the diagnostics are not enabled by default.

+++ src/internal/diagnostics.h
@@ -139,7 +139,7 @@ Diagnostics diagnostics{};
         {                                                                                          \
             diagnostics.tt_move_is_best.update(*scratchpad_.tt_move_score >= BEST_RESULT.score);   \
         }                                                                                          \
-        diagnostics.num_moves_considered.update(static_cast<u64>(MOVE_COUNT));                     \
+        diagnostics.num_moves_considered.update(MOVE_COUNT);                                       \
     } while (false)
 
 #define DIAGNOSTICS_UPDATE_AFTER_MOVE(TYPE, SCORE)                                                 \
